+++ clients/admin-ui/src/features/datastore-connections/types.ts
@@ -1,10 +1,10 @@
+import { ConnectionType, SystemType } from "~/types/api";
+
 import {
   AccessLevel,
   ConnectionTestStatus,
-  ConnectionType,
   DisabledStatus,
   SaasType,
-  SystemType,
   TestingStatus,
 } from "./constants";
 
@@ -163,7 +163,7 @@ export const isDatastoreConnection = (obj: any): obj is DatastoreConnection =>
 
 export type DatastoreConnectionParams = {
   search: string;
-  connection_type?: ConnectionType[];
+  connection_type?: string[];
   test_status?: TestingStatus;
   system_type?: SystemType;
   disabled_status?: DisabledStatus;